makefileification

+++ Test.cpp
@@ -12,7 +12,7 @@ void test()
 
     v1 = sort(v1);
 
-    for(int i=1;i<v1.size();i++)
+    for(int i=1;i<(int)v1.size();i++)
         if(v1[i]<v1[i-1])
             is_sorted = false;
